Refactor and fix dsUocSo() and dsPhanTuChung()
